requirements.txt
opencv-python
elevenlabs
captacity
openai
pydub
numpy
openai-whisper
ffmpeg
opencv-python
python-dotenv